Add level reset
Pressing `C-r` to reset the sprites, commands and colliders gives us a
nice quick feedback loops for testing.

Need to figure out a neat way of letting the player know they can do
this.

Unfortunately having the reset function also reset the key-pressed
listeners will cause a cyclical dependency, should be fine to not need
this, just good to keep in mind.