Adding proper platform collision
Finishing off the first 5 levels required having horizontal platform
collisions. This needed a bit of a revamp of how quip applies collide
functions after detecting a collision.

There also needed to be distinctinos made between each of the sides of
a platform that can be hit, this allows us to properly adjust
velocity, set `:landed` flags, and correct position to stop us getting
stuck in corners.