Stop players leaving world bounds
Easiest solution was to just define a thin platrform for each side of
the screen which levels can include in their sprite list.

Also fixed a bug where a double dash in mid air could get you to clip
into a platoform. Was only hapenning when vel-x was greater than half
the width of the player so the check o the players position being
outside the platform was failing. Now we offset the position we're
checking against by the appropriate velocity dimension.