Adding level goals and transitions
Now each level has a goal for the player to try and get to.

On collision with the goal sprite we set it's animation and then check
for this in the level's update-fn to see if we've completed the
level. This isn't the nicest way of doing this, but it'll probably
take us as far as we need this time.

Currently using the default fade-to-black level transition, would be
nice to upgrade this to something fancier.