Add play animations
Adding animations for idle, jump, dash and turn, also figured out a
simple and not too horrible way of getting the sprite to do one loop
of it's animation and then revet to idle.

Must make sure we update the :animation-timer values if we change the
number of frames in, or speed of, an animation.