Remove commands from credits scene before game is played.
It lokos weird to have the list of commands there if you don't know
what they are. So now clicking the play button on the menu is what
popultaes the commands.

This means our menu scene is now referencing the credits ns, which
isn't ideal, but as long as we don't keep doing things like this we
should be fine.