Bundle required fonts with game
We're using Ubuntu Mono as the font for the commands, unsurprisingly
this isnt installed by default on Windows or Mac, so we're bundling
the .ttf file with the game.

Making sure to create the font sduring setup and store it on the main
state so we can access it everywhere without the calling the costly
create-font each time.