Add effects when collecting new commands
The code involved is getting a little busy, but we now have a homing
particle group created when the pickup sprite finishes it's animation,
the homing group targets the location we expect the command to appear
in (they're alphabetised right? will this cause issues?), the command
is created with a display delay so it only appears when we expect to
particles to reach it, we can be reasonably loose about the timing
here.

We're also playing the new command sound on the frame where the
comamnds display-delay is zero, which is a little cheecky, but
whatevs.

Quip could really use the ability to create delayed actions.

Also tweens, but that's probably a lot of work.