Add reset as command
In levels where you can die you now hit a hidden pickup using a custom
collision function that checks for you falling below the screen, this
pickup adds the reset command to the scene.

Also made it so that pickup particles are yellow so they're more
visible and also made it so they don't collide with the platforms
using a :phasing? flag, a little hacky, but it works.

I'm gonna keep the ctrl+r reset as it's useful in the levels where you
can't die.