Add intro scene
Getting the user to understand typing commands with as little input as
possible.

As a happy coincidence the music seems to sync up with the animation
really nicely, but this might not be the case on slower machines.

The code required to get this working is pretty nasty. Definitely
running into limitations of quip, can't keep hacking stuff in like
this.

Closing in on something submittable here :)